Extract IOContext
