+++ src/main/java/com/github/nianna/karedi/KarediApp.java
@@ -113,7 +113,7 @@ public class KarediApp extends Application {
 	}
 
 	public void exit(Event event) {
-		if (appContext.needsSaving()) {
+		if (appContext.ioContext.needsSaving()) {
 			if (!saveChangesIfUserWantsTo()) {
 				event.consume();
 				return;
@@ -148,8 +148,8 @@ public class KarediApp extends Application {
 	}
 
 	public boolean saveChangesIfUserWantsTo() {
-		if (appContext.needsSaving()) {
-			Alert alert = new SaveChangesAlert(getFileName(appContext.getActiveFile()));
+		if (appContext.ioContext.needsSaving()) {
+			Alert alert = new SaveChangesAlert(getFileName(appContext.ioContext.getActiveFile()));
 
 			Optional<ButtonType> result = alert.showAndWait();
 			if (!result.isPresent()) {
@@ -160,7 +160,7 @@ public class KarediApp extends Application {
 				}
 				if (result.get() == SaveChangesAlert.SAVE_BUTTON) {
 					appContext.actionContext.execute(KarediActions.SAVE);
-					if (appContext.needsSaving()) {
+					if (appContext.ioContext.needsSaving()) {
 						// Save failed or was cancelled by the user
 						return false;
 					}
