+++ src/main/java/com/github/nianna/karedi/context/AppContext.java
@@ -1,36 +1,18 @@
 package com.github.nianna.karedi.context;
 
-import com.github.nianna.karedi.I18N;
 import com.github.nianna.karedi.KarediApp;
 import com.github.nianna.karedi.KarediApp.ViewMode;
-import com.github.nianna.karedi.song.Note;
-import com.github.nianna.karedi.song.Song;
-import com.github.nianna.karedi.song.SongTrack;
-import com.github.nianna.karedi.song.tag.Tag;
-import com.github.nianna.karedi.song.tag.TagKey;
-import com.github.nianna.karedi.txt.TxtFacade;
-import javafx.beans.binding.BooleanBinding;
 import javafx.beans.property.ReadOnlyObjectProperty;
 import javafx.beans.property.ReadOnlyObjectWrapper;
-import javafx.collections.ObservableList;
 
-import java.io.File;
-import java.util.List;
 import java.util.logging.Logger;
 
 public class AppContext {
 
 	public static final Logger LOGGER = Logger.getLogger(KarediApp.class.getPackage().getName());
 
-	private final ReadOnlyObjectWrapper<File> activeFile = new ReadOnlyObjectWrapper<>();
-
 	private final ReadOnlyObjectWrapper<ViewMode> activeViewMode = new ReadOnlyObjectWrapper<>(
 			KarediApp.getInstance().getViewMode());
-
-	private final TxtFacade txtFacade = new TxtFacade();
-
-	// Convenience bindings for actions
-	public final BooleanBinding activeFileIsNull = activeFileProperty().isNull();
 
 	public final ActiveSongContext activeSongContext = new ActiveSongContext();
 
@@ -49,61 +31,12 @@ public class AppContext {
 
 	public final CommandContext commandContext = new CommandContext(this);
 
+	public final IOContext ioContext = new IOContext(activeSongContext, audioContext, commandContext);
+
 	public final ActionContext actionContext = new ActionContext(this);
 
 	public AppContext() {
 		LOGGER.setUseParentHandlers(false);
-	}
-
-	public ReadOnlyObjectProperty<File> activeFileProperty() {
-		return activeFile.getReadOnlyProperty();
-	}
-
-	public File getActiveFile() {
-		return activeFileProperty().get();
-	}
-
-	public void setActiveFile(File file) {
-		this.activeFile.set(file);
-		KarediApp.getInstance().updatePrimaryStageTitle(file);
-	}
-
-	public void loadSongFile(File file) {
-		loadSongFile(file, true);
-	}
-
-	public void loadSongFile(File file, boolean resetPlayer) {
-		if (file != null) {
-			reset(resetPlayer);
-			setActiveFile(file);
-			Song song = txtFacade.loadFromTxtFile(file);
-			song.getTagValue(TagKey.MP3)
-					.ifPresent(audioFileName -> audioContext.loadAudioFile(new File(file.getParent(), audioFileName)));
-			activeSongContext.setSong(song);
-			LOGGER.info(I18N.get("load.success"));
-		}
-	}
-
-	public boolean saveSongToFile(File file) {
-		if (file != null) {
-			if (txtFacade.saveSongToFile(file, activeSongContext.getSong())) {
-				commandContext.updateLastSavedCommand();
-				return true;
-			}
-		}
-		return false;
-	}
-
-	public void exportSongToFile(File file, ObservableList<Tag> tags, List<SongTrack> tracks) {
-		txtFacade.exportSongToFile(file, tags, tracks);
-	}
-
-	public Song loadFromClipboard() {
-		return txtFacade.loadFromClipboard();
-	}
-
-	public void saveToClipboard(List<Note> selectedNotes) {
-		txtFacade.saveToClipboard(selectedNotes);
 	}
 
 	public ReadOnlyObjectProperty<ViewMode> activeViewModeProperty() {
@@ -118,19 +51,8 @@ public class AppContext {
 		activeViewMode.set(viewMode);
 	}
 
-	public boolean needsSaving() {
-		return activeSongContext.getSong() != null && commandContext.hasUnsavedCommands();
-	}
-
 	public Logger getMainLogger() {
 		return LOGGER;
-	}
-
-	public void reset(boolean resetPlayer) {
-		activeSongContext.setSong(null);
-		if (resetPlayer) {
-			audioContext.reset();
-		}
 	}
 
 }
