+++ src/main/java/com/github/nianna/karedi/context/actions/ClipboardAction.java
@@ -10,6 +10,6 @@ abstract class ClipboardAction extends ContextfulKarediAction {
     }
 
     protected Song buildSongFromClipboard() {
-        return appContext.loadFromClipboard();
+        return ioContext.loadFromClipboard();
     }
 }
